Expand containing folders when a source is selected (#1793)
* Expand containing folders when a source is selected

* Address CC

* Address CR
