+++ app/components/SourceSelector.vue.ts
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import { Component } from 'vue-property-decorator';
+import { Component, Watch } from 'vue-property-decorator';
 import { Inject } from '../services/core/injector';
 import { SourcesService } from 'services/sources';
 import { ScenesService, ISceneItemNode, TSceneNode } from 'services/scenes';
@@ -215,6 +215,17 @@ export default class SourceSelector extends Vue {
   canShowActions(sceneNodeId: string) {
     const node = this.scene.getNode(sceneNodeId);
     return node.isItem() || node.getNestedItems().length;
+  }
+
+  get lastSelectedId() {
+    return this.selectionService.state.lastSelectedId;
+  }
+
+  @Watch('lastSelectedId')
+  expandSelectedFolders() {
+    const node = this.scenesService.activeScene.getNode(this.lastSelectedId);
+    if (!node) return;
+    this.expandedFoldersIds = node.getPath().slice(0, -1);
   }
 
   get activeItemIds() {
